Move quote service object into enclave handle classes
